🚚 Remove non-Redux files from the data folder
